+++ plugins/auth-backend/src/lib/flow/authFlowHelpers.test.ts
@@ -62,9 +62,11 @@ describe('oauth helpers', () => {
       const encoded = safelyEncodeURIComponent(JSON.stringify(data));
 
       postMessageResponse(mockResponse, appOrigin, data);
-      expect(mockResponse.setHeader).toBeCalledTimes(3);
-      expect(mockResponse.end).toBeCalledTimes(1);
-      expect(mockResponse.end).toBeCalledWith(expect.stringContaining(encoded));
+      expect(mockResponse.setHeader).toHaveBeenCalledTimes(3);
+      expect(mockResponse.end).toHaveBeenCalledTimes(1);
+      expect(mockResponse.end).toHaveBeenCalledWith(
+        expect.stringContaining(encoded),
+      );
     });
 
     it('should post a message back with payload error', () => {
@@ -80,9 +82,11 @@ describe('oauth helpers', () => {
       const encoded = safelyEncodeURIComponent(JSON.stringify(data));
 
       postMessageResponse(mockResponse, appOrigin, data);
-      expect(mockResponse.setHeader).toBeCalledTimes(3);
-      expect(mockResponse.end).toBeCalledTimes(1);
-      expect(mockResponse.end).toBeCalledWith(expect.stringContaining(encoded));
+      expect(mockResponse.setHeader).toHaveBeenCalledTimes(3);
+      expect(mockResponse.end).toHaveBeenCalledTimes(1);
+      expect(mockResponse.end).toHaveBeenCalledWith(
+        expect.stringContaining(encoded),
+      );
     });
 
     it('should call postMessage twice but only one of them with target *', () => {
@@ -166,9 +170,9 @@ describe('oauth helpers', () => {
       };
 
       postMessageResponse(mockResponse, appOrigin, data);
-      expect(mockResponse.setHeader).toBeCalledTimes(3);
-      expect(mockResponse.end).toBeCalledTimes(1);
-      expect(mockResponse.end).toBeCalledWith(
+      expect(mockResponse.setHeader).toHaveBeenCalledTimes(3);
+      expect(mockResponse.end).toHaveBeenCalledTimes(1);
+      expect(mockResponse.end).toHaveBeenCalledWith(
         expect.stringContaining('Adam%20l%27H%C3%B4pital'),
       );
     });