+++ plugins/catalog-backend-module-github/src/providers/GitHubEntityProvider.ts
@@ -36,9 +36,9 @@ import { Logger } from 'winston';
 import {
   readProviderConfigs,
   GitHubEntityProviderConfig,
-  TopicTypes,
 } from './GitHubEntityProviderConfig';
 import { getOrganizationRepositories, Repository } from '../lib/github';
+import { satisfiesTopicFilter } from '../lib/util';
 
 /**
  * Discovers catalog files located in [GitHub](https://github.com).
@@ -64,23 +64,24 @@ export class GitHubEntityProvider implements EntityProvider {
     },
   ): GitHubEntityProvider[] {
     const integrations = ScmIntegrations.fromConfig(config);
-    const integration = integrations.github.byHost('github.com');
 
-    if (!integration) {
-      throw new Error(
-        `There is no GitHub config that matches github. Please add a configuration entry for it under integrations.github`,
+    return readProviderConfigs(config).map(providerConfig => {
+      const integrationHost = providerConfig.host;
+      const integration = integrations.github.byHost(integrationHost);
+
+      if (!integration) {
+        throw new Error(
+          `There is no GitHub config that matches host ${integrationHost}. Please add a configuration entry for it under integrations.github`,
+        );
+      }
+
+      return new GitHubEntityProvider(
+        providerConfig,
+        integration,
+        options.logger,
+        options.schedule,
       );
-    }
-
-    return readProviderConfigs(config).map(
-      providerConfig =>
-        new GitHubEntityProvider(
-          providerConfig,
-          integration,
-          options.logger,
-          options.schedule,
-        ),
-    );
+    });
   }
 
   private constructor(
@@ -184,30 +185,16 @@ export class GitHubEntityProvider implements EntityProvider {
 
   private matchesFilters(repositories: Repository[]) {
     const repositoryFilter = this.config.filters?.repository;
-    const topicFilter = this.config.filters?.topic;
+    const topicFilters = this.config.filters?.topic;
 
     const matchingRepositories = repositories.filter(r => {
       const repoTopics: string[] = r.repositoryTopics.nodes.map(
         node => node.topic.name,
       );
-      const satisfiesTopicFilter = (
-        topics: string[],
-        filter: typeof topicFilter,
-      ) => {
-        if (!filter || !filter.name) return true;
-        if (filter.type === TopicTypes.Includes && topics.includes(filter.name))
-          return true;
-        if (
-          filter.type === TopicTypes.Excludes &&
-          !topics.includes(filter.name)
-        )
-          return true;
-        return false;
-      };
       return (
         !r.isArchived &&
         (!repositoryFilter || repositoryFilter.test(r.name)) &&
-        satisfiesTopicFilter(repoTopics, topicFilter) &&
+        satisfiesTopicFilter(repoTopics, topicFilters) &&
         r.defaultBranchRef?.name
       );
     });