+++ plugins/cloudbuild/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-cloudbuild
 
+## 0.3.9-next.1
+
+### Patch Changes
+
+- 817f3196f6: Updated React Router dependencies to be peer dependencies.
+- 763fb81e82: Internal refactor to use more type safe code when dealing with route parameters.
+- d669d89206: Minor API signatures cleanup
+- Updated dependencies
+  - @backstage/core-components@0.11.1-next.1
+  - @backstage/core-plugin-api@1.0.6-next.1
+  - @backstage/plugin-catalog-react@1.1.4-next.1
+
 ## 0.3.9-next.0
 
 ### Patch Changes