+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-cost-insights
 
+## 0.11.31-next.1
+
+### Patch Changes
+
+- 817f3196f6: Updated React Router dependencies to be peer dependencies.
+- d669d89206: Minor API signatures cleanup
+- Updated dependencies
+  - @backstage/core-components@0.11.1-next.1
+  - @backstage/core-plugin-api@1.0.6-next.1
+
 ## 0.11.31-next.0
 
 ### Patch Changes